backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyJWT==2.8.0
Werkzeug==3.0.1
python-dateutil==2.8.2
